Revert "Merge branch 'master' of https://bitbucket.org/UO258092/ips-l7-g1"
This reverts commit 4efb1732a31fd504c9d3a8abd6b378593756aa66, reversing
changes made to bcd696478009730289fde06d169f94caba4942e9.

+++ Logic.Db/Dto/CompetitionDto.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 
 namespace Logic.Db.Dto {
     public class CompetitionDto {
-        public long ID;
+        public int ID;
         public string Name;
         public TypeCompetition Type;
         public double Km;
